fix(pods): Fix --pod argument
When a pod name is specified using the --pod argument, we take it as a string, but never attempt to
resolve a pod based on this name. The code which connects to the specific pod assumes a JSON object
with the pod metadata, and so the connection fails. This change will attempt to resolve the
pod_name, and provides an error message if that pod_name does not exist.